Add files via upload
The CSS I made for the site and the HTML template. I also used the bootstrap CSS as well with the HTML. This is just a prototype of form the UI of the site. I'm unsure how to upload the folder holding the example I used in the picture I sent in group me, but I think the template should be self-explanatory.